fix redundant env var loading
I have added the env var loader as well as tests. The loader will not override any system environment variables, so system env vars are prioritized over the ones passed into the loader. This is consistent with python-dotenv's methods.

Currently tests cover that env vars in a .env file that is passed to the autoloader will load the correct values with the correct names. The tests do not ensure that correct env vars from the system are loaded in. The test also does not cover whether multiple configs can be loaded together using this method, but those tests should be covered elsewhere and I do not expect the env loader to interact with multiple configs in a breaking manner.