Feature/dotenv loading (#42)
* add env loader and configure auto loader to accept .env

* add tests for env parsing

* fix redundant env var loading

I have added the env var loader as well as tests. The loader will not override any system environment variables, so system env vars are prioritized over the ones passed into the loader. This is consistent with python-dotenv's methods.

Currently tests cover that env vars in a .env file that is passed to the autoloader will load the correct values with the correct names. The tests do not ensure that correct env vars from the system are loaded in. The test also does not cover whether multiple configs can be loaded together using this method, but those tests should be covered elsewhere and I do not expect the env loader to interact with multiple configs in a breaking manner.

* fix too many spaces in typehint

* fix return type hint to be str str

* remove unused import

* make enumerate more pythonic

* remove unused raw_config import

* add os.environ adding env var to test, add docstring

* remove redundant if check

A filepath must always be given and will be validated by the config.register_config function. Therefore no check is needed that the filepath or file exists.